network: Wait for network device in UpdateInterface
A new field PciAddr has been added to grpc.Interface.
This is the PCI identifier of the network device.

In order to support network hotplug, we need to wait for the
device to be online first based on the PCI address of the device
inside the VM before updating its configuration.

Fixes #305

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>